Move handlers into handlers.js in prep for move to serviceFacades
